+++ LeetCodeNet/G0601_0700/S0637_average_of_levels_in_binary_tree/Solution.cs
@@ -37,7 +37,7 @@ public class Solution {
         }
         if (level < count.Count) {
             count[level]++;
-            avg[level] += node.val;
+            avg[level] += (double) node.val;
         } else {
             count.Add(1);
             avg.Add(node.val);
